Improves randomness of items on market (#263)
* Previous market used a fixed offset which resulted in group of items always appearing together in the market
* This update provides a unique market seed and offset to improve randomness of market

+++ contracts/market/src/constants.cairo
@@ -1,5 +1,5 @@
 // number of items per level per survivor
-const NUMBER_OF_ITEMS_PER_LEVEL: u128 = 20;
+const NUMBER_OF_ITEMS_PER_LEVEL: u8 = 20;
 
 // number of loot items
 const NUM_LOOT_ITEMS: u64 = 101;
